requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
seaborn>=0.12.0
matplotlib>=3.7.0
python-dotenv>=1.0.0 